python: simplify access to Magika's API results
This patch simplifies how to access Magika's result. In its current version, the
result could be accessed with `res.value.output` or `res.value.dl`. The extra
`.value` is needed because the current API returns `StatusOr[MagikaResult]`
instead of `MagikaResult`. Returning a `StatusOr` is a clean and battle-tested
way of implement API that could potentially fail.  However, in the vast majority
of cases, the API does not (and in some cases, cannot) fail: the
`identify_paths` API would only fail in case of permission issues; the
`identify_bytes` cannot fail (as in: there is no code path that would return a
non-OK status).

This patch allows clients to access the underlying result without needing to go
through `.value`: e.g., `res.output` or `res.dl`. This is very similar to what
we supported for versions earlier than this one.

This change comes at a cost: instead of having a generic `StatusOr`, we now have
`StatusOrMagikaResult`. While it would be possible to implement "access the
underlying value" with the generic `StatusOr[T]` (for example by using
`__getattr__` and forwarding calls to the underlying `.value`), this would mess
up type resolution. Since such `StatusOr[T]` or `StatusOrMagikaResult` is hidden
from external clients, we accept to make our code less generic in exchange of
making the most common case more usable.