python: do not alter warnings' filter
We recently added a couple of `DeprecationWarning`s, and to make sure users would see them, we also set `warnings.simplefilter` to `always`. However, this inadvertently modified the warnings' settings for all other modules as well. This is not nice.

This patch removes the edit to the simplefilter (going back to default) and uses `stacklevel`s `warn` argument to increase the `DeprecationWarning` visibiilty. This is not a silver bullet as if one uses the deprecated property "three levels or more" deep, the warning is not visible.

Very recent versions of python added the `skip_file_prefixes` argument to `warn`, but it's not worth using it and break compatibility with older versions of python.