python: most functions are now around Seekable, removed a lot of duplicate code
There were a number of methods specialized for path, bytes, and stream variants. Most of them were unnecessary, are now replaced with the same implementation around the Seekable interface.

The Seekable object has also been simplified: it's not a concrete class, and we got rid of the various subclasses (such as File, Buffer, and Stream). The functions are just written about Seekable now.