js: restrict input type to Uint8Array and Buffer
As far as I can tell, supporting Uint16Array is not very useful, and it adds complexity (as in: it's something we need to convert from/to, something we need to test for, etc.), and it can even be a bit ambiguous (e.g., is one supposed to pass "two bytes at the same time" in each Uint16Array element"? This is not so far fetched as that's how some wide code points are dealt with in Windows => that's something we need to explain/document).