gh: avoid overtriggerig js import scenario
In GH actions, the "path" filter is not used for tags push. This means that any tag push would trigger this gh action, even if the push has nothing to do with it.

Fixed by triggering only for branch pushes to main (note: the branch filter is used for tags as well), and for any PR that relates to this.